+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntitySearchField.java
@@ -911,8 +911,8 @@ public final class EntitySearchField extends HintTextField {
 							new SwingEntityTableModel(searchField.model.entityDefinition().type(),
 											emptyList(), searchField.model.connectionProvider());
 
-			FilterTable<Entity, Attribute<?>> filterTable = FilterTable.builder(tableModel,
-											entityTableColumns(tableModel.entityDefinition()))
+			FilterTable<Entity, Attribute<?>> filterTable = FilterTable.builder(tableModel)
+							.columns(entityTableColumns(tableModel.entityDefinition()))
 							.autoResizeMode(JTable.AUTO_RESIZE_ALL_COLUMNS)
 							.cellRendererFactory(EntityTableCellRenderer.factory())
 							.selectionMode(searchField.singleSelection() ?